tools/google_drive/requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
dify-plugin>=0.3.0
